Move getEmailParams to feedbackUtils
