+++ pages/api/feedback-rc.ts
@@ -1,50 +1,14 @@
 import type { NextApiRequest, NextApiResponse } from "next"
 import aws from "aws-sdk"
-import { type SendEmailRequest } from "aws-sdk/clients/ses"
-
-import { appConfig } from "../../src/config/config"
-import {
-  getFeedbackEmailText,
-  getFeedbackEmailHTML,
-} from "../../src/utils/feedbackUtils"
-import { encodeHTML } from "../../src/utils/appUtils"
+import { getEmailParams } from "../../src/utils/feedbackUtils"
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === "POST") {
     if (!req.body) res.status(400).json({ error: "Malformed request" })
     aws.config.update({ region: "us-east-1" })
 
-    const fields = JSON.parse(req.body)
-
-    const fullUrl = encodeHTML(req.headers.referer)
-
-    const emailText = getFeedbackEmailText(fullUrl, fields)
-    const emailHTML = getFeedbackEmailHTML(fullUrl, fields)
-
-    const emailParams: SendEmailRequest = {
-      Destination: {
-        ToAddresses: [appConfig.libAnswersEmail],
-      },
-      Message: {
-        Body: {
-          Html: {
-            Charset: "UTF-8",
-            Data: emailHTML,
-          },
-          Text: {
-            Charset: "UTF-8",
-            Data: emailText,
-          },
-        },
-        Subject: {
-          Charset: "UTF-8",
-          Data: "SCC Feedback",
-        },
-      },
-      Source: appConfig.sourceEmail,
-      ReplyToAddresses: [fields.email || appConfig.sourceEmail],
-    }
-
+    const emailParams = getEmailParams(req.body, req.headers.referer)
+    console.log(emailParams)
     // Create the promise and SES service object
     const sendPromise = new aws.SES({ apiVersion: "2010-12-01" })
       .sendEmail(emailParams)
